requirements.txt
anthropic==0.86.0
fastapi==0.135.2
openai==2.29.0
pydantic==2.12.5
python-dotenv==1.2.2
qdrant-client==1.17.1
tqdm==4.67.3
uvicorn==0.42.0